refactor: move some manifest constants in solution to appstudio/constants.ts (#1697)
fix: solution
