+++ packages/fx-core/src/plugins/solution/fx-solution/constants.ts
@@ -2,13 +2,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import {
-  IBot,
-  IComposeExtension,
-  IConfigurableTab,
-  IStaticTab,
-  UserError,
-} from "@microsoft/teamsfx-api";
+import { UserError } from "@microsoft/teamsfx-api";
 
 /**
  * Void is used to construct Result<Void, FxError>.
@@ -39,15 +33,6 @@ export const PERMISSION_REQUEST = "permissionRequest";
  * Config key whose value is either javascript, typescript or csharp.
  */
 export const PROGRAMMING_LANGUAGE = "programmingLanguage";
-
-/**
- * Config keys that are useful for generating remote teams app manifest
- */
-export const REMOTE_MANIFEST = "manifest.source.json";
-export const FRONTEND_ENDPOINT = "endpoint";
-export const FRONTEND_DOMAIN = "domain";
-export const BOT_ID = "botId";
-export const LOCAL_BOT_ID = "localBotId";
 
 export const DEFAULT_PERMISSION_REQUEST = [
   {
@@ -136,159 +121,6 @@ export const LOCAL_DEBUG_TEAMS_APP_ID = "localDebugTeamsAppId";
 // Teams App Id for remote
 export const REMOTE_TEAMS_APP_ID = "remoteTeamsAppId";
 
-export const TEAMS_APP_MANIFEST_TEMPLATE = `{
-    "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.9/MicrosoftTeams.schema.json",
-    "manifestVersion": "1.9",
-    "version": "{version}",
-    "id": "{appid}",
-    "packageName": "com.microsoft.teams.extension",
-    "developer": {
-        "name": "Teams App, Inc.",
-        "websiteUrl": "{baseUrl}",
-        "privacyUrl": "{baseUrl}/index.html#/privacy",
-        "termsOfUseUrl": "{baseUrl}/index.html#/termsofuse"
-    },
-    "icons": {
-        "color": "color.png",
-        "outline": "outline.png"
-    },
-    "name": {
-        "short": "{appName}",
-        "full": "This field is not used"
-    },
-    "description": {
-        "short": "Short description of {appName}.",
-        "full": "Full description of {appName}."
-    },
-    "accentColor": "#FFFFFF",
-    "bots": [],
-    "composeExtensions": [],
-    "configurableTabs": [],
-    "staticTabs": [],
-    "permissions": [
-        "identity",
-        "messageTeamMembers"
-    ],
-    "validDomains": [],
-    "webApplicationInfo": {
-        "id": "{appClientId}",
-        "resource": "{webApplicationInfoResource}"
-    }
-}`;
-
-export const COMPOSE_EXTENSIONS_TPL: IComposeExtension[] = [
-  {
-    botId: "{botId}",
-    commands: [
-      {
-        id: "createCard",
-        context: ["compose"],
-        description: "Command to run action to create a Card from Compose Box",
-        title: "Create Card",
-        type: "action",
-        parameters: [
-          {
-            name: "title",
-            title: "Card title",
-            description: "Title for the card",
-            inputType: "text",
-          },
-          {
-            name: "subTitle",
-            title: "Subtitle",
-            description: "Subtitle for the card",
-            inputType: "text",
-          },
-          {
-            name: "text",
-            title: "Text",
-            description: "Text for the card",
-            inputType: "textarea",
-          },
-        ],
-      },
-      {
-        id: "shareMessage",
-        context: ["message"],
-        description: "Test command to run action on message context (message sharing)",
-        title: "Share Message",
-        type: "action",
-        parameters: [
-          {
-            name: "includeImage",
-            title: "Include Image",
-            description: "Include image in Hero Card",
-            inputType: "toggle",
-          },
-        ],
-      },
-      {
-        id: "searchQuery",
-        context: ["compose", "commandBox"],
-        description: "Test command to run query",
-        title: "Search",
-        type: "query",
-        parameters: [
-          {
-            name: "searchQuery",
-            title: "Search Query",
-            description: "Your search query",
-            inputType: "text",
-          },
-        ],
-      },
-    ],
-    messageHandlers: [
-      {
-        type: "link",
-        value: {
-          domains: ["*.botframework.com"],
-        },
-      },
-    ],
-  },
-];
-export const BOTS_TPL: IBot[] = [
-  {
-    botId: "{botId}",
-    scopes: ["personal", "team", "groupchat"],
-    supportsFiles: false,
-    isNotificationOnly: false,
-    commandLists: [
-      {
-        scopes: ["personal", "team", "groupchat"],
-        commands: [
-          {
-            title: "intro",
-            description: "Send introduction card of this Bot",
-          },
-          {
-            title: "show",
-            description: "Show user profile by calling Microsoft Graph API with SSO",
-          },
-        ],
-      },
-    ],
-  },
-];
-export const CONFIGURABLE_TABS_TPL: IConfigurableTab[] = [
-  {
-    configurationUrl: "{baseUrl}/index.html#/config",
-    canUpdateConfiguration: true,
-    scopes: ["team", "groupchat"],
-  },
-];
-
-export const STATIC_TABS_TPL: IStaticTab[] = [
-  {
-    entityId: "index",
-    name: "Personal Tab",
-    contentUrl: "{baseUrl}/index.html#/tab",
-    websiteUrl: "{baseUrl}/index.html#/tab",
-    scopes: ["personal"],
-  },
-];
-
 export const DoProvisionFirstError = new UserError(
   "DoProvisionFirst",
   "DoProvisionFirst",
@@ -297,12 +129,6 @@ export const DoProvisionFirstError = new UserError(
 export const CancelError = new UserError("UserCancel", "UserCancel", "Solution");
 // This is the max length specified in
 // https://developer.microsoft.com/en-us/json-schemas/teams/v1.7/MicrosoftTeams.schema.json
-export const TEAMS_APP_SHORT_NAME_MAX_LENGTH = 30;
-
-// Default values for the developer fields in manifest.
-export const DEFAULT_DEVELOPER_WEBSITE_URL = "https://www.example.com";
-export const DEFAULT_DEVELOPER_TERM_OF_USE_URL = "https://www.example.com/termofuse";
-export const DEFAULT_DEVELOPER_PRIVACY_URL = "https://www.example.com/privacy";
 
 export enum SolutionTelemetryEvent {
   CreateStart = "create-start",
